test: add metadata unit tests (#45)
* Add test repository data

This change adds repository data to use in unit testing, keeping
it the same as in python-tuf

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Add file system simulator

This change adds a mock for os functions

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Add test client simulator

This change implements client simulator for working with local
repositories

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Add metadata unit tests

This change adds unit tests for part of the metadata methods

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Add metadata api tests

This change implements api unit tests for metadata with the same
logic as in python-tuf

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Fix go linter in test simulators

This change fixes go linter error of not checking errors returned
by methods in the file system and test client simulators

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Create clean test metadata repository

As the `signature` package allows loading keys from RSA PEM files
only, the keys and roles from the python-tuf implementation are
not suitable for the current implementation. This is why a clean
repo is created with new keys in the appropriate format

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Add metadata sign/verify test

This change adds test coverage for signing and verigying metadata
with a pre-set keys

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Cleanup unnecessary functionality

This change removes unused keys and test utils for the metadata
unit tests. The removed test utils include filesystem simulator,
which is replaced by the setup functionality and some tidy-up
around the changes

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Fix metadata api tests to verify error type

This change updates error assertions in the metadata api tests
to compare error types rather than only a message

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Improve metadata tests

This change provides improvements to the metadata api tests

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

---------

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>
Co-authored-by: Ivana Atanasova <iyovcheva@iyovcheva7MD6R.vmware.com>