Add updater tests (#65)
* Add repository simulator for testing updater

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Fix updater loadRoot

Previously the version was iterating as leq to the upperbound which
was causing errors and is incorrect. For comparison in the analogue
python-tuf implementation `range()` works as only `<` to the upper
bound

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Fix metadata IsDelegatedPath

Previously it was not always correclty checked. This change fixes
the path pattern check to be more presise

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Fix updater preOrderDepthFirstWalk

There was a bug in preOrderDepthFirstWalk to always break a loop
for a specific use-case, without `terminating` to be set to true

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Add updater tests

This change adds updater top level update and consistent snapshot
tests, using the repository simulator

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Fix failing tests

This change provides a fix for the failing CI due to logging
updates, as well as updater changes in main

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Fix linter issues

This change fixes go linter issues in the updater tests

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

* Fix updater tests: elements order mismatch

When testing if client fetches and stores delegated metadata files
with the correct version prefix when ConsistentSnapshot is true it
sometimes happens that the implementation of GetTargetInfo allows
differences in the order the roles are recorded in the FetchTraker.
This change fixes the test to not look into the order

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>

---------

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>